mobends update

+++ src/minecraft/net/minecraft/entity/Entity.java
@@ -1931,11 +1931,11 @@ public abstract class Entity implements ICommandSender, Cullable
 
         if (set)
         {
-            this.dataWatcher.updateObject(0, Byte.valueOf((byte)(b0 | 1 << flag)));
+            this.dataWatcher.updateObject(0, (byte) (b0 | 1 << flag));
         }
         else
         {
-            this.dataWatcher.updateObject(0, Byte.valueOf((byte)(b0 & ~(1 << flag))));
+            this.dataWatcher.updateObject(0, (byte) (b0 & ~(1 << flag)));
         }
     }
 
@@ -1946,7 +1946,7 @@ public abstract class Entity implements ICommandSender, Cullable
 
     public void setAir(int air)
     {
-        this.dataWatcher.updateObject(1, Short.valueOf((short)air));
+        this.dataWatcher.updateObject(1, (short) air);
     }
 
     public void onStruckByLightning(EntityLightningBolt lightningBolt)
